Replace float with CGFloat in NS(Attributed)String+Geometrics. Add generics to NSArray(OEAdditions).

+++ OpenEmu/NS(Attributed)String+Geometrics.h
@@ -128,31 +128,31 @@ extern NSInteger gNSStringGeometricsTypesetterBehavior ;
 @interface NSAttributedString (Geometrics) 
 
 // Measuring Attributed Strings
-- (NSSize)sizeForWidth:(float)width 
-				height:(float)height ;
-- (float)heightForWidth:(float)width ;
-- (float)widthForHeight:(float)height ;
+- (NSSize)sizeForWidth:(CGFloat)width
+				height:(CGFloat)height ;
+- (CGFloat)heightForWidth:(CGFloat)width ;
+- (CGFloat)widthForHeight:(CGFloat)height ;
 
 @end
 
 @interface NSString (Geometrics)
 
 // Measuring a String With Attributes
-- (NSSize)sizeForWidth:(float)width 
-				height:(float)height
+- (NSSize)sizeForWidth:(CGFloat)width
+				height:(CGFloat)height
 			attributes:(NSDictionary*)attributes ;
-- (float)heightForWidth:(float)width
-			 attributes:(NSDictionary*)attributes ;
-- (float)widthForHeight:(float)height
-			 attributes:(NSDictionary*)attributes ;
+- (CGFloat)heightForWidth:(CGFloat)width
+               attributes:(NSDictionary*)attributes ;
+- (CGFloat)widthForHeight:(CGFloat)height
+               attributes:(NSDictionary*)attributes ;
 
 // Measuring a String with a constant Font
-- (NSSize)sizeForWidth:(float)width 
-				height:(float)height
+- (NSSize)sizeForWidth:(CGFloat)width
+				height:(CGFloat)height
 				  font:(NSFont*)font ;
-- (float)heightForWidth:(float)width
-				   font:(NSFont*)font ;
-- (float)widthForHeight:(float)height
-				   font:(NSFont*)font ;
+- (CGFloat)heightForWidth:(CGFloat)width
+                     font:(NSFont*)font ;
+- (CGFloat)widthForHeight:(CGFloat)height
+                     font:(NSFont*)font ;
 
 @end
